posix: properly fix size mismatch of keyboard/joystick_cfg arrays
fixes #563

unlike #566, this also fixes the gtk2 frontend automatically.

+++ desmume/src/frontend/posix/shared/ctrlssdl.h
@@ -66,7 +66,7 @@ extern const char *key_names[NB_KEYS];
 /* Current keyboard configuration */
 extern u32 keyboard_cfg[NB_KEYS];
 /* Current joypad configuration */
-extern u16 joypad_cfg[NB_KEYS];
+extern u32 joypad_cfg[NB_KEYS];
 /* Number of detected joypads */
 extern u16 nbr_joy;
